Added AES128 encryption to the Arduino end of the communication.
The protocol now uses a 16byte string, where the first two characters
are the unique identifier for the lock, followed by 12 characters of
the card UID, followed by padding (which is based on the millis() value
from the Arduino).
The padding should mean that even if a card is read repeatedly, the
data should be different each time.