Updated dorbot_server to use AES128 encryption. It's not yet fully functional, but is nearly there (in testing, it seems that the radio doesn't receive the data from the Arduino component).
The encryption is carried out at the Arduino end in AES128 and then broadcast to the RPI.
The RPi decrypts the packet, which contains 2 characters to ID the sender (so it's theoretically possible to have multiple doors with different access permissions on each controlled from one server)
The next 12 characters is the key UID in  2 byte pieces, separated by spaces
Finally, there are 3 characters of padding derived from the Arduino's clock. THese change each time the packet is sent and reduce the similarity between packets. Because the attacker can't then know exactly what the contents of the packet are (e.g. by reading the code on github), it's hopefully even more difficult to break the encryption password.