Renamed the file server.cpp to dorbot_server.cp
dorbot_server.cpp
=================

It is now possible to have encrypted communication between the RPi and the arduino on the door. AES128 is used to encrypt the transmission to try to prevent folks from hacking the system.

TODO
----

Add config file for the DB details (at the moment they're baked in); encryption key; and other config information

makefile
========

Makefile now exists to make the build easier.